Add option to use docker-compose for the Postgres db (locally)
Note: as part of this, I've revamped the `database.yml` – please see the comments in there for details.

Also: this change now adds the `.env` file into source control - since dotenv is only used during development and test, this won't affect production environments. Later on, we can add `.env.test` (committed to repo) and `.env.local` (NOT committed to repo) files for overrides and secrets via env vars. The rationale for all this is to have a better "out of the box" setup with default config (via env vars) in place.