Subscription: users must now specify a timezone for their schedule
In the database, we still store the data in normalised UTC – we convert back and forth between the user's timezone and UTC so the user only sees the schedule in their selected timezone.

BREAKING CHANGE: we now only allow users to select **1** time slot – this makes adjusting the schedule based on the user's timezone possible (with multiple time slots there are cases where it's impossible to convert to/from the user's timezones, due to the days adjustment [forward or backwards]).

KNOWN ISSUE: this doesn't handle daylight savings at all :( The worst case is that users receive emails an hour before their actual scheduled time slot (note: that since schedules are shown to users as two hours, but processed every hour, the case where daylight savings would result in it being sent one hour forward would still be okay as it's still within the advertised time slot).