Move the languages selector from the header to the courses index page
Since the only translated content are the courses, it makes sense to only show switching of languages on the courses index page.

Note: we still make use of the site wide I18n handling… so if/when we do start adding site-wide translations then it would make sense to move this language selector back into the header.